Port the examples and rewrite the weight matrix tests
The 19 example and demo files called initialize_gpu_context and
release_gpu_resources, which went with the global backend; the backend is a
member of each engine now, constructed and released with it, so the calls
just go away.

weight_matrix_tests.cpp tested a class that no longer exists in that shape --
260 references to per-edge planes, exact mode, refit and Sk round-trips.
Deleted rather than patched, and replaced with tests against what the class
actually is now: edge numbering by real out-degree, exact reconstruction of
projection weights, tiny weights surviving the basis at 1e-12, delays
round-tripping as exact integers, prototype lookup through the run table, and
an assertion that nothing allocated scales with node_count *
max_neighbor_count -- a regression there is the whole design coming undone
and would be invisible in every numerical test.

Two bugs the new tests caught:

  - load_from_disk restored the basis but not the delay fast path, so a
    loaded matrix answered the default one tick for every edge. The flag is
    state, not something derivable from U/V, so it has to be written.

  - Emitting the synapse integration before the synapse body meant the
    conductance-based rejection never ran, and an unsupported synapse failed
    with the wrong reason. Both emitters read the same type, so whichever
    ran first decided the diagnostic; the rejection now happens once up
    front, where the emission order cannot change it.

144 tests pass. The six that fail are one root cause, reported separately:
networks whose delays are drawn i.i.d. per edge fragment into one projection
run per edge, and the basis then needs a rank equal to the edge count.

+++ tests/nml_tests.cpp
@@ -1035,7 +1035,7 @@ TEST(Export, adjacency_projection_matches_the_stored_edges) {
 
     // The plain source -> targets form WeightMatrix consumes is derived, not stored, so
     // there is only one edge list that can be wrong.
-    Vector<Vector<s64>> adjacency = build_adjacency_list(result);
+    Vector<Vector<s32>> adjacency = build_adjacency_list(result);
 
     ASSERT_EQ(adjacency.size(), result.neurons.size());
     for (usize source = 0; source < result.neurons.size(); source += 1) {
@@ -1047,8 +1047,8 @@ TEST(Export, adjacency_projection_matches_the_stored_edges) {
         }
     }
 
-    EXPECT_EQ(adjacency[0], (Vector<s64>{4}));
-    EXPECT_EQ(adjacency[2], (Vector<s64>{3}));
+    EXPECT_EQ(adjacency[0], (Vector<s32>{4}));
+    EXPECT_EQ(adjacency[2], (Vector<s32>{3}));
     EXPECT_TRUE(adjacency[1].empty());
 }
